Notes freq as float

+++ examples/examples-stk/streams-stk-desktop/streams-stk-desktop.ino
@@ -22,7 +22,7 @@ StreamCopy copier(out, in);
 MusicalNotes notes;
 
 float note_amplitude = 0.5;
-static uint16_t notes_array[] = { // frequencies aleatoric C-scale
+static float notes_array[] = { // frequencies aleatoric C-scale
     N_C3, N_D3, N_E3, N_F3, N_G3, N_A3, N_B3,
     N_C4, N_D4, N_E4, N_F4, N_G4, N_A4, N_B4,
     N_C5, N_D5, N_E5, N_F5, N_G5, N_A5, N_B5
